refactor: consolidate trace component
