+++ traces_component.go
@@ -1,15 +1,136 @@
 package emqutiti
 
 import (
+	"fmt"
+	"io"
 	"os"
+	"strings"
 	"time"
 
 	"github.com/charmbracelet/bubbles/list"
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/x/ansi"
 
 	"github.com/marang/emqutiti/ui"
 )
+
+// traceItem represents a single trace configuration and its runtime tracer.
+type traceItem struct {
+	key    string
+	cfg    TracerConfig
+	tracer *Tracer
+	counts map[string]int
+	loaded bool
+}
+
+func (t *traceItem) FilterValue() string { return t.key }
+func (t *traceItem) Title() string       { return t.key }
+func (t *traceItem) Description() string {
+	status := "stopped"
+	if t.tracer != nil {
+		if t.tracer.Running() {
+			status = "running"
+		} else if t.tracer.Planned() {
+			status = "planned"
+		}
+	} else if time.Now().Before(t.cfg.Start) {
+		status = "planned"
+	}
+	var parts []string
+	counts := t.counts
+	if t.tracer != nil {
+		counts = t.tracer.Counts()
+	} else if !t.loaded {
+		if c, err := tracerLoadCounts(t.cfg.Profile, t.cfg.Key, t.cfg.Topics); err == nil {
+			t.counts = c
+			t.loaded = true
+			counts = c
+		}
+	}
+	for _, tp := range t.cfg.Topics {
+		parts = append(parts, fmt.Sprintf("%s:%d", tp, counts[tp]))
+	}
+	if len(parts) > 0 {
+		status += " " + strings.Join(parts, " ")
+	}
+	var times []string
+	if !t.cfg.Start.IsZero() {
+		times = append(times, t.cfg.Start.Format(time.RFC3339))
+	}
+	if !t.cfg.End.IsZero() {
+		times = append(times, t.cfg.End.Format(time.RFC3339))
+	}
+	if len(times) > 0 {
+		status += " " + strings.Join(times, " -> ")
+	}
+	return status
+}
+
+// tracesState groups state related to tracing functionality.
+type tracesState struct {
+	list    list.Model
+	items   []*traceItem
+	form    *traceForm
+	view    list.Model
+	viewKey string
+}
+
+// traceMsgItem holds a trace message with its sequence number.
+type traceMsgItem struct {
+	idx int
+	msg TracerMessage
+}
+
+// FilterValue implements list.Item for filtering.
+func (t traceMsgItem) FilterValue() string { return t.msg.Payload }
+
+// Title implements list.Item.
+func (t traceMsgItem) Title() string { return t.msg.Topic }
+
+// Description implements list.Item.
+func (t traceMsgItem) Description() string { return t.msg.Payload }
+
+// traceMsgDelegate renders trace messages with numbering and timestamp.
+type traceMsgDelegate struct{ m *model }
+
+// Height returns the line height for an item.
+func (d traceMsgDelegate) Height() int { return 2 }
+
+// Spacing satisfies list.ItemDelegate.
+func (d traceMsgDelegate) Spacing() int { return 0 }
+
+// Update is a no-op for this delegate.
+func (d traceMsgDelegate) Update(msg tea.Msg, m *list.Model) tea.Cmd { return nil }
+
+// Render draws a trace message item with numbering and timestamp.
+func (d traceMsgDelegate) Render(w io.Writer, m list.Model, index int, item list.Item) {
+	it := item.(traceMsgItem)
+	width := m.Width()
+	innerWidth := width - 2
+	if innerWidth < 0 {
+		innerWidth = 0
+	}
+	ts := it.msg.Timestamp.Format("2006-01-02T15:04:05.000000000Z07:00")
+	header := fmt.Sprintf("%d %s %s:", it.idx, ts, it.msg.Topic)
+	lines := []string{lipgloss.PlaceHorizontal(innerWidth, lipgloss.Left,
+		lipgloss.NewStyle().Foreground(ui.ColBlue).Render(header))}
+	for _, l := range strings.Split(it.msg.Payload, "\n") {
+		wrapped := ansi.Wrap(l, innerWidth, " ")
+		for _, wl := range strings.Split(wrapped, "\n") {
+			lines = append(lines, lipgloss.PlaceHorizontal(innerWidth, lipgloss.Left,
+				lipgloss.NewStyle().Foreground(ui.ColSub).Render(wl)))
+		}
+	}
+	barColor := ui.ColDarkGray
+	if index == d.m.traces.view.Index() {
+		barColor = ui.ColPurple
+	}
+	bar := lipgloss.NewStyle().Foreground(barColor)
+	lines = ui.FormatHistoryLines(lines, width, bar)
+	fmt.Fprint(w, strings.Join(lines, "\n"))
+}
 
 type traceTickMsg struct{}
